refactor(js_interop_gen): decompose transformer.dart by extracting NamespaceFlattener
