+++ js_interop_gen/lib/src/interop_gen/transform/transformer.dart
@@ -17,7 +17,6 @@ import '../../ast/types.dart';
 import '../../banned_names.dart';
 import '../../js/annotations.dart';
 import '../../js/filesystem_api.dart';
-import '../../js/helpers.dart';
 import '../../js/typescript.dart' as ts;
 import '../../js/typescript.types.dart';
 import '../../translator.dart';
@@ -27,6 +26,7 @@ import '../qualified_name.dart';
 import '../transform.dart';
 import 'dependency_walker.dart';
 import 'export_reference.dart';
+import 'namespace_flattener.dart';
 import 'utils.dart';
 
 /// A class for transforming nodes in a given [file]
@@ -86,6 +86,14 @@ class Transformer {
        namer = UniqueNamer(),
        _fileName = file,
        assert(_sourceFile != null || file != null, 'Source file must be known');
+
+  Documentation? parseAndTransformDocumentation(TSNamedDeclaration node) =>
+      _parseAndTransformDocumentation(node);
+
+  EnumDeclaration transformEnum(
+    TSEnumDeclaration enumeration, {
+    UniqueNamer? namer,
+  }) => _transformEnum(enumeration, namer: namer);
 
   // TODO(nikeokoronkwo): Handle default exports
   /// Transforms a TypeScript AST Node [TSNode] into a Dart representable [Node]
@@ -239,163 +247,9 @@ class Transformer {
     UniqueNamer? namer,
     NamespaceDeclaration? parent,
   }) {
-    namer ??= this.namer;
-
-    final namespaceName = (namespace.name as TSIdentifier).text;
-
-    // get modifiers
-    final modifiers = namespace.modifiers?.toDart ?? [];
-    final isExported = modifiers.any((m) {
-      return m.kind == TSSyntaxKind.ExportKeyword;
-    });
-
-    final currentNamespaces = parent != null
-        ? parent.namespaceDeclarations.where((n) => n.name == namespaceName)
-        : nodeMap.findByName(namespaceName).whereType<NamespaceDeclaration>();
-
-    final (name: dartName, :id) = currentNamespaces.isEmpty
-        ? namer.makeUnique(namespaceName, 'namespace')
-        : (name: null, id: null);
-
-    final scopedNamer = ScopedUniqueNamer();
-
-    final outputNamespace = currentNamespaces.isNotEmpty
-        ? currentNamespaces.first
-        : NamespaceDeclaration(
-            name: namespaceName,
-            dartName: dartName,
-            id: id!,
-            exported: isExported,
-            topLevelDeclarations: {},
-            namespaceDeclarations: {},
-            nestableDeclarations: {},
-            documentation: _parseAndTransformDocumentation(namespace),
-          );
-
-    // TODO: We can implement this in classes and interfaces.
-    //  however, since namespaces and modules are a thing,
-    //  let's keep that in mind
-    /// Updates the state of the given declaration,
-    /// allowing cross-references between types and declarations in the
-    /// namespace, including the namespace itself
-    void updateNSInParent() {
-      if (parent != null) {
-        if (currentNamespaces.isNotEmpty ||
-            parent.namespaceDeclarations.any((n) => n.name == namespaceName)) {
-          parent.namespaceDeclarations.remove(currentNamespaces.first);
-          parent.namespaceDeclarations.add(outputNamespace);
-        } else {
-          outputNamespace.parent = parent;
-          parent.namespaceDeclarations.add(outputNamespace);
-        }
-      } else {
-        nodeMap.update(
-          outputNamespace.id.toString(),
-          (v) => outputNamespace,
-          ifAbsent: () => outputNamespace,
-        );
-      }
-    }
-
-    void transformDeclAndAppendParent(
-      NamespaceDeclaration outputNamespace,
-      TSNode decl,
-    ) {
-      if (outputNamespace.nodes.contains(decl)) return;
-      if (decl.kind == TSSyntaxKind.EnumMember) {
-        final tsEnum = (decl as TSEnumMember).parent;
-        // parse whole enum
-        final transformedEnum = _transformEnum(tsEnum, namer: namer);
-
-        // add enum
-        if (parent != null) {
-          parent.nestableDeclarations.add(transformedEnum);
-          parent.nodes.add(tsEnum);
-        } else {
-          nodes.add(tsEnum);
-          nodeMap.add(transformedEnum);
-        }
-
-        // add all members to namespace
-        outputNamespace.nodes.addAll(tsEnum.members.toDart);
-      } else {
-        final outputDecls = transformAndReturn(
-          decl,
-          namer: scopedNamer,
-          parent: outputNamespace,
-        );
-        switch (decl.kind) {
-          case TSSyntaxKind.ClassDeclaration ||
-              TSSyntaxKind.InterfaceDeclaration:
-            final outputDecl = outputDecls.single as TypeDeclaration;
-            outputDecl.parent = outputNamespace;
-            outputNamespace.nestableDeclarations.add(outputDecl);
-          case TSSyntaxKind.EnumDeclaration:
-            final outputDecl = outputDecls.single as EnumDeclaration;
-            outputDecl.parent = outputNamespace;
-            outputNamespace.nestableDeclarations.add(outputDecl);
-          case TSSyntaxKind.TypeAliasDeclaration:
-            final outputDecl = outputDecls.single as TypeAliasDeclaration;
-            outputDecl.parent = outputNamespace;
-            outputNamespace.nestableDeclarations.add(outputDecl);
-          default:
-            outputNamespace.topLevelDeclarations.addAll(outputDecls);
-        }
-        outputNamespace.nodes.add(decl);
-      }
-
-      // update namespace state
-      updateNSInParent();
-    }
-
-    // preload nodemap
-    updateNSInParent();
-
-    // to reduce probing, we can use exported decls instead
-    final symbol = typeChecker.getSymbolAtLocation(namespace.name);
-    final exports = symbol?.exports?.toDart;
-
-    if (exports case final exportedMap?) {
-      for (final symbol in exportedMap.values) {
-        final decls = symbol.getDeclarations()?.toDart ?? [];
-        try {
-          final aliasedSymbol = typeChecker.getAliasedSymbol(symbol);
-          decls.addAll(aliasedSymbol.getDeclarations()?.toDart ?? []);
-        } catch (_) {
-          // throws error if no aliased symbol, so ignore
-        }
-        for (final decl in decls) {
-          // TODO: We could also ignore namespace decls with the same name, as
-          //  a single instance should consider such non-necessary
-          transformDeclAndAppendParent(outputNamespace, decl);
-        }
-      }
-      // fallback
-    } else {
-      if (namespace.body case final namespaceBody?
-          when namespaceBody.kind == TSSyntaxKind.ModuleBlock) {
-        for (final statement
-            in (namespaceBody as TSModuleBlock).statements.toDart) {
-          transformDeclAndAppendParent(outputNamespace, statement);
-        }
-      } else if (namespace.body case final namespaceBody?) {
-        // namespace import
-        _transformNamespace(
-          namespaceBody as TSNamespaceDeclaration,
-          namer: scopedNamer,
-          parent: outputNamespace,
-        );
-      }
-    }
-
-    // final update on namespace state
-    updateNSInParent();
-
-    // index names
-    namer.markUsedSet(scopedNamer);
-
-    // get the exported symbols from the namespace
-    return outputNamespace;
+    return NamespaceFlattener(
+      this,
+    ).transformNamespace(namespace, namer: namer, parent: parent);
   }
 
   /// Transforms a TS Class or Interface declaration into a node representing
